fix nested env configs, and clean up sampling
