+++ rl/pufferlib/play.py
@@ -7,7 +7,7 @@ from agent.policy_store import PolicyStore
 
 def play(cfg: OmegaConf, policy_store: PolicyStore):
     device = cfg.device
-    vecenv = make_vecenv(cfg.env, cfg.vectorization, num_envs=1, render_mode="human")
+    vecenv = make_vecenv(cfg.eval.env, cfg.vectorization, num_envs=1, render_mode="human")
 
     obs, _ = vecenv.reset()
     env = vecenv.envs[0]
