+++ rl/pufferlib/vecenv.py
@@ -5,13 +5,10 @@ import pufferlib
 import pufferlib.utils
 import pufferlib.vector
 import hydra
-from util.runtime_configuration import setup_omega_conf
+from mettagrid.mettagrid_env import MettaGridEnv
 
 def make_env_func(cfg: DictConfig, buf=None, render_mode='rgb_array'):
-    setup_omega_conf()
-    env_cfg = OmegaConf.create(cfg)
-    env = hydra.utils.instantiate(env_cfg, env_cfg, buf=buf, render_mode=render_mode, _recursive_=False)
-    return env
+    return MettaGridEnv(cfg, buf=buf, render_mode=render_mode)
 
 def make_vecenv(
     env_cfg: OmegaConf,
