+++ util/runtime_configuration.py
@@ -18,64 +18,6 @@ def seed_everything(seed, torch_deterministic):
         torch.manual_seed(seed)
     torch.backends.cudnn.deterministic = torch_deterministic
 
-def uniform(min_val, max_val, center, *, _root_):
-    sampling = _root_.get("sampling", 0)
-    if sampling == 0:
-        return center
-    else:
-
-        center = (max_val + min_val) // 2
-        # Calculate the available range on both sides of the center
-        left_range = center - min_val
-        right_range = max_val - center
-
-        # Scale the ranges based on the sampling parameter
-        scaled_left = min(left_range, sampling * left_range)
-        scaled_right = min(right_range, sampling * right_range)
-
-        # Generate a random value within the scaled range
-        val = np.random.uniform(center - scaled_left, center + scaled_right)
-
-        # Clip to ensure we stay within [min_val, max_val]
-        val = np.clip(val, min_val, max_val)
-
-        # Return integer if the original values were integers
-        return int(round(val)) if isinstance(center, int) else val
-
-def choose(*args):
-    return random.choice(args)
-
-def div(a, b):
-    return a // b
-
-def sub(a, b):
-    return a - b
-
-def make_odd(a):
-    return max(3, a // 2 * 2 + 1)
-
-def sweep_space(space, min_val, max_val, center=None, *, _root_):
-    result = {
-        "space": space,
-        "min": min_val,
-        "max": max_val,
-        "search_center": center,
-    }
-    if space == "int":
-        result["is_int"] = True
-        result["space"] = "linear"
-    return OmegaConf.create(result)
-
-
-def setup_omega_conf():
-    OmegaConf.register_new_resolver("div", div, replace=True)
-    OmegaConf.register_new_resolver("uniform", uniform, replace=True)
-    OmegaConf.register_new_resolver("sub", sub, replace=True)
-    OmegaConf.register_new_resolver("make_odd", make_odd, replace=True)
-    OmegaConf.register_new_resolver("choose", choose, replace=True)
-
-    OmegaConf.register_new_resolver("ss", sweep_space, replace=True)
-
 def setup_metta_environment(cfg):
     # Set environment variables to run without display
     os.environ['GLFW_PLATFORM'] = 'osmesa'  # Use OSMesa as the GLFW backend
@@ -89,7 +31,6 @@ def setup_metta_environment(cfg):
     warnings.filterwarnings('ignore', category=DeprecationWarning, module='pkg_resources')
     warnings.filterwarnings('ignore', category=DeprecationWarning, module='pygame.pkgdata')
 
-    setup_omega_conf()
     if cfg.dist_cfg_path is not None:
         dist_cfg = OmegaConf.load(cfg.dist_cfg_path)
         cfg.run = dist_cfg.run
